requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# OCR Dependencies
pytesseract==0.3.10
Pillow==10.1.0

# Data Processing
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2